requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2

# NLP and ML
transformers==4.36.0
torch==2.1.1
sentence-transformers==2.2.2
spacy==3.7.2
nltk==3.8.1

# Vector database and search
chromadb==0.4.18
faiss-cpu==1.7.4
numpy==1.24.3

# Email processing
email-validator==2.1.0
imaplib2==3.6
email-mime-multipart==1.0.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
aiofiles==23.2.0
pandas==2.1.4
tqdm==4.66.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
